+++ e2e/tests/mlb/methods_off_payment.spec.js
@@ -3,7 +3,7 @@ import { addProductToCart } from "../../flows/add_to_cart";
 import goToCheckout from "../../flows/go_to_checkout";
 import fillGuestEmail from "../../flows/fill_guest_email";
 import fillShippingForm from "../../flows/fill_shipping_form";
-import { fillDocumentDataPt } from '../../flows/fill_off_document';
+import { fillDocumentData } from '../../flows/fill_off_document';
 import { base, mlb } from "../../data/stores";
 
 test('test create payment with Boleto', async ({ page }) => {
@@ -11,26 +11,10 @@ test('test create payment with Boleto', async ({ page }) => {
   await goToCheckout(page, mlb.storeUrl);
   await fillGuestEmail(page);
   await fillShippingForm(page);
-  await fillDocumentDataPt(page, process.env.CPF);
+  await fillDocumentData(page, process.env.CPF);
   await page.locator('#payment_methods_off').first().check();
   await page.getByRole('button', { name: 'Next: Place Order' }).click();
-  await expect(page.locator('#checkout-main')).toContainText('Via boleto em Boleto');
+  await expect(page.locator('#checkout-main')).toContainText('Via invoice at Boleto');
   await page.getByRole('button', { name: 'Place Order' }).click();
-  await expect(page.locator('#maincontent')).toContainText('Onde pagar: Boleto');
-});
-
-test('test create payment with Pagamento na lotérica sem boleto', async ({ page }) => {
-  await addProductToCart(page, base.storeUrl);
-  await goToCheckout(page, mlb.storeUrl);
-  await fillGuestEmail(page);
-  await fillShippingForm(page);
-  await fillDocumentDataPt(page, process.env.CPF);
-  await page.locator('li').filter({ hasText: 'Pagamento na lotérica sem' }).locator('#payment_methods_off').check();
-  await page.getByRole('button', { name: 'Next: Place Order' }).click();
-  await expect(page.locator('#checkout-main')).toContainText('Via boleto em Pagamento na lotérica sem boleto');
-  await page.getByRole('button', { name: 'Place Order' }).click();
-  await expect(page.locator('#maincontent')).toContainText('Onde pagar: Pagamento na lotérica sem boleto');
-  await page.getByRole('button', { name: 'Abrir boleto' }).click();
-  await expect(page.locator('#mercadopago-client')).toContainText('Pague R$ 44,09 na lotérica mais próxima.');
-
+  await expect(page.locator('#maincontent')).toContainText('Where to pay: Boleto');
 });