+++ e2e/tests/mlu/methods_off_payment.spec.js
@@ -3,7 +3,7 @@ import { addProductToCart } from "../../flows/add_to_cart";
 import goToCheckout from "../../flows/go_to_checkout";
 import fillGuestEmail from "../../flows/fill_guest_email";
 import fillShippingForm from "../../flows/fill_shipping_form";
-import { fillDocumentDataEs } from '../../flows/fill_off_document';
+import { fillDocumentData } from '../../flows/fill_off_document';
 import { base, mlu } from "../../data/stores";
 
 test('test create payment with Abitab', async ({ page }) => {
@@ -11,7 +11,7 @@ test('test create payment with Abitab', async ({ page }) => {
   await goToCheckout(page, mlu.storeUrl);
   await fillGuestEmail(page);
   await fillShippingForm(page);
-  await fillDocumentDataEs(page, process.env.DOC_OUTRO);
+  await fillDocumentData(page, process.env.DOC_OUTRO);
   await page.locator('li').filter({ hasText: 'Abitab' }).locator('#payment_methods_off').check();
   await page.getByRole('button', { name: 'Next: Place Order' }).click();
   await page.waitForTimeout(1000);
@@ -26,7 +26,7 @@ test('test create payment with Redpagos', async ({ page }) => {
   await goToCheckout(page, mlu.storeUrl);
   await fillGuestEmail(page);
   await fillShippingForm(page);
-  await fillDocumentDataEs(page, process.env.DOC_OUTRO);
+  await fillDocumentData(page, process.env.DOC_OUTRO);
   await page.locator('li').filter({ hasText: 'Redpagos' }).locator('#payment_methods_off').check();
   await page.getByRole('button', { name: 'Next: Place Order' }).click();
   await page.waitForTimeout(1000);